requirements.txt
Flask==2.3.2
Flask-MySQLdb==2.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
Flask-CORS==4.0.0
python-dotenv==1.0.0
user-agents==2.2.0
Werkzeug==2.3.6
PyMySQL==1.1.0
cryptography==41.0.1
